Separate Datadog.Trace.ClrProfiler.Managed.Loader startup behavior into separate files for .NET Framework and .NET Core.
